Execute selection immediately if there is a complete match.
